requirements.txt
streamlit 
ultralytics 
opencv-python-headless 
pillow 
imagehash 
scikit-image